% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities-data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Geographical Locations of World Cities}
\format{
A data frame with 43645 rows and 5 variables.
\describe{
  \item{name}{name of the city}
  \item{country}{recorded values for entity 1, in US dollars.}
  \item{population}{recorded values for entity 2, in US dollars.}
  \item{lat}{recorded values for entity 3, in US dollars.}
  \item{long}{recorded values for entity 4, in US dollars.}
}
}
\source{
\url{https://cran.r-project.org/package=maps}
}
\usage{
data(cities)
}
\description{
Data set with geographical locations of cities. These data can also be found in the \code{maps} package.
}
\examples{
data(cities)
}
\keyword{datasets}
