% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview.R
\name{cityview}
\alias{cityview}
\title{Create a City View}
\usage{
cityview(
  name = NULL,
  zoom = 1,
  theme = c(
    "vintage", "modern", "bright", "delftware", "comic",
    "rouge", "original", "midearth", "batik", "vice"
  ),
  border = c(
    "none", "circle", "rhombus", "square",
    "hexagon", "octagon", "decagon", "bbox"
  ),
  halftone = NULL,
  legend = FALSE,
  places = 0,
  license = TRUE,
  timeout = 25,
  filename = NULL,
  verbose = TRUE,
  cache = TRUE,
  bot = FALSE
)
}
\arguments{
\item{name}{a character specifying the name of the city as provided by
\code{list_cities()}, an object created using
\code{new_city()}, or a row of the ouput of
\code{list_cities()}. If \code{NULL} (default), chooses a
random city.}

\item{zoom}{a numeric value specifying the amount of zoom. Values > 1
increase zoom and speed up computation time, while values < 1
decrease zoom and increase computation time. For zoom levels
below 0.5, the computation time can be very long.}

\item{theme}{a character specifying the theme of the plot, or a named list
specifying a custom theme (see the details section for more
information about the composition of this list). Possible
pre-specified themes are \code{vintage} (default),
\code{modern}, \code{bright}, \code{delftware}, \code{comic},
\code{rouge}, \code{original}, \code{midearth}, \code{batik}
and \code{vice}.}

\item{border}{a character specifying the type of border to use. Possible
options are \code{none} (default), \code{circle},
\code{rhombus}, \code{square}, \code{hexagon} (6 vertices),
\code{octagon} (8 vertices), \code{decagon} (10 vertices) and
\code{bbox} (the bounding box for the entire city, argument
\code{zoom} will be ignored).}

\item{halftone}{a character specifying the color of applied halftone dither.}

\item{legend}{logical. Whether to add a distance measurer and a compass in
the bottom left corner of the image.}

\item{places}{an integer specifying how many suburb, quarter and
neighbourhood names to add to the image.}

\item{license}{logical. Whether to add the OpenStreetMap licence in the
bottom right corner of the figure.}

\item{timeout}{a value specifying the timeout (in seconds) for the Overpass
server. It may be necessary to increase this value for large
or populated areas because the server may time out before all
data are delivered (if this occurs you will receive the
following error message: \code{runtime error: Query timed out
in "recurse" at line ... after ... seconds}.}

\item{filename}{character. If specified, the function exports the plot at an
appropriate size and does not return a \code{ggplot2} object.}

\item{verbose}{logical. Whether to show a progress bar during execution.}

\item{cache}{logical. Whether to cache the data for the image so that they
do not need to be requested again when calling the function
with a different theme.}

\item{bot}{logical. Enable functionality used by the X bot.}
}
\description{
Create a city view showcasing a particular city or region using
  OpenStreetMap (OSM) data retreived trough the Overpass API. Please note
  that OpenStreetMap is open data and can be freely utilized for any purpose,
  as long as proper credit is given to OpenStreetMap and its contributors.
}
\details{
The \code{theme} argument can take a custom list as input (see the
  example). This list must contain all of the following elements:

\code{colors}
\itemize{
 \item{\code{background}:  A single color to be used for the background.}
 \item{\code{water}:       A single color to be used for the water.}
 \item{\code{landuse}:     A single color or a vector of multiple colors to
                           be used for the landuse.}
 \item{\code{contours}:    A single color to be used for the contours (lines)
                           of landuse and buildings.}
 \item{\code{streets}:     A single color to be used for the streets.}
 \item{\code{rails}:       A single color or a vector of two colors to be
                           used for the rails.}
 \item{\code{buildings}:   A single color or a vector of multiple colors to
                           be used for the buildings.}
 \item{\code{text}:        A single color to be used for the text.}
}
\code{font}
\itemize{
 \item{\code{family}:      A string specifying the family of the font.}
 \item{\code{face}:        A string specifying the face of the font.}
 \item{\code{scale}:       A single value specifying the expansion factor of
                           the characters in the font.}
 \item{\code{append}:      Optional. A string to append the city name with at
                           both sides.}
}
\code{size}
\itemize{
 \item{\code{borders}:    A named list containing sizes for the borders
                          \code{contours}, \code{water}, \code{canal} and
                          \code{river}.}
 \item{\code{streets}:    A named list contianing sizes for the streets
                          \code{path}, \code{residential}, \code{structure},
                          \code{tertiary}, \code{secondary}, \code{primary},
                          \code{motorway}, \code{rails} and \code{runway}.}
}
}
\examples{
\dontrun{
# 1. Simple example
# Create a city view of Amsterdam in a circle
cityview(name = "Amsterdam", border = "circle")

# 2. Advanced example
# Custom theme (black, beige and white), streets only, direct export
myTheme <- list(
  colors = list(
    background = "#232323",
    water = "#232323",
    landuse = "#232323",
    contours = "#232323",
    streets = "#d7b174",
    rails = c("#d7b174", "#232323"),
    buildings = "#232323",
    text = "#ffffff",
    waterlines = "#232323"
  ),
  font = list(
    family = "serif",
    face = "bold",
    scale = 1,
    append = "\u2014"
  ),
  size = list(
    borders = list(
      contours = 0.15,
      water = 0.4,
      canal = 0.5,
      river = 0.6
    ),
    streets = list(
      path = 0.2,
      residential = 0.3,
      structure = 0.35,
      tertiary = 0.4,
      secondary = 0.5,
      primary = 0.6,
      motorway = 0.8,
      rails = 0.75,
      runway = 3
    )
  )
)
cityview(
  name = "Amsterdam", theme = myTheme,
  border = "square", filename = "Amsterdam.png"
)
}
}
\references{
\url{https://www.openstreetmap.org}
}
\seealso{
\code{\link{list_cities}}
         \code{\link{cityview_shiny}}
         \code{\link{new_city}}
         \code{\link{city_themes}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
