% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview_shiny.R
\name{cityview_shiny}
\alias{cityview_shiny}
\title{[Shiny] Create an Aerial City View}
\usage{
cityview_shiny(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{logical. If \code{FALSE} (the default), the app will run in the viewer panel. If \code{TRUE}, the app will run in the system's default web browser.}
}
\description{
Create an aerial city view in Shiny.
}
\examples{
\dontrun{
# Start the shiny app
cityview_shiny()
}
}
\seealso{
\code{\link{list_cities}} \code{\link{cityview}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
\keyword{shiny}
