% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_raw}
\alias{geocode_raw}
\title{Geocode a Location Using Various Methods}
\usage{
geocode_raw(location, method = "osm")
}
\arguments{
\item{location}{A character string specifying the location name to geocode. It must be provided as a single string.}

\item{method}{A character string specifying the geocoding method to use. Defaults to "osm" (OpenStreetMap). Supported methods include:
\code{"osm"}, \code{"census"}, \code{"arcgis"}, \code{"census_simple"}, \code{"geocodio"}, \code{"mapbox"},
\code{"google"}, \code{"bing"}, \code{"here"}, \code{"tomtom"}, \code{"nominatim"}, and \code{"tiger"}.}
}
\value{
A data frame containing the following columns:
  \item{name}{The name of the city or location.}
  \item{country}{The country in which the location is situated, inferred from the display name.}
  \item{lat}{The latitude of the location.}
  \item{long}{The longitude of the location.}
  \item{population}{The population of the location (not available in this function, returns \code{NA}).}
}
\description{
This function takes a location name as a string and returns the latitude and longitude coordinates using various geocoding methods.
  It supports multiple geocoding services and handles API keys where necessary.
}
\details{
This function attempts to geocode the specified location name using the chosen method. If the method requires an API key,
  it will check for the necessary environment variable and return an error if the key is missing. The function outputs a data frame
  containing the geocoded coordinates and location information.
}
\examples{
\dontrun{
  # Geocode a location using OpenStreetMap (default)
  geocode_raw("New York, USA")

  # Geocode a location using Google (API key required)
  geocode_raw("Tokyo, Japan", method = "google")
}

}
