% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cities.R
\name{list_cities}
\alias{list_cities}
\title{List or Search City Names}
\usage{
list_cities(match = NULL)
}
\arguments{
\item{match}{string to match.}
}
\description{
List all city names.
}
\examples{
# List all cities containing "Ams"
list_cities(match = "Ams")
}
\seealso{
\code{\link{cityview}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{search}
