% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_city.R
\name{new_city}
\alias{new_city}
\title{Specify a New City}
\usage{
new_city(name = NULL, country = NULL, lat = NULL, long = NULL)
}
\arguments{
\item{name}{string to be printed as the city name.}

\item{country}{string to be printed as the country.}

\item{lat}{numeric value to be used as the latitude.}

\item{long}{numeric value to be used as the longitude.}
}
\value{
a data frame containing the new city alongside its respective
  country and coordinates.
}
\description{
Define a city using country, latitude and longitude coordinates.
  The returned object can be used a input for \code{name} argument the
  \code{cityview()} function.
}
\examples{
city <- new_city(
  name = "Lagos", country = "Portugal",
  lat = 37.10, long = -8.68
)
\dontrun{
cityview(name = city)
}
}
\seealso{
\code{\link{cityview}} \code{\link{cityview_shiny}}
  \code{\link{list_cities}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{new}
