% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcityviews-methods.R
\name{rcityviews-methods}
\alias{rcityviews-methods}
\alias{print.rcityviewsTheme}
\title{Methods for rcityviews objects}
\usage{
\method{print}{rcityviewsTheme}(x, ...)
}
\arguments{
\item{...}{further arguments, currently ignored.}

\item{object, x}{an object of class \code{rcityviewsTheme}.}
}
\value{
The \code{print} methods simply print and return nothing.
}
\description{
Methods defined for objects returned from the \code{\link{city_themes}}
  functions.
}
