\name{rcityviews-package}
\alias{rcityviews}
\alias{rcityviews-package}
\docType{package}
\title{R City Views}
\description{
  \code{rcityviews} is an homage to the programming language \code{R},
  open-source geographic data and the art of map making. It provides code and
  examples to render customizable stylized city maps using data from
  OpenStreetMap.
  
  For documentation on \code{rcityviews} itself, including the manual and user
  guide for the package, worked examples, and other tutorial information visit
  the  \href{https://koenderks.github.io/rcityviews/}{package website}.
}

\author{
  \tabular{ll}{
    Koen Derks (maintainer, author)\tab <koen-derks@hotmail.com> \cr
  }
 
  Please use the citation provided by \R when citing this package. 
  A BibTex entry is available from \code{citation("rcityviews")}.
}

\seealso{
Useful links:
\itemize{
  \item The \href{https://x.com/rcityviews}{tx feed} to check the regularly city views.
  \item The \href{https://github.com/koenderks/rcityviews/issues}{issue page} to submit a bug report or feature request.
}
}

\keyword{rcityviews} 
\keyword{package}
