% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview.R
\name{cityview}
\alias{cityview}
\title{Create an Aerial City View}
\usage{
cityview(name = NULL,
         zoom = 1,
         theme = c(
           "vintage", "colored", "delftware", "lichtenstein",
           "original", "rouge", "modern"
         ),
         border = c(
           "none", "circle", "rhombus", "square",
           "hexagon", "octagon", "decagon"
         ),
         halftone = c("none", "light", "modern"),
         places = 0,
         legend = FALSE,
         filename = NULL,
         verbose = TRUE,
         license = TRUE,
         bot = FALSE)
}
\arguments{
\item{name}{a character specifying the name of the city as provided by \code{list_cities()}. If \code{NULL} (default), chooses a random city.}

\item{zoom}{a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.}

\item{theme}{a character specifying the theme of the plot, or a named list specifying a custom theme (see the details section for more information about the composition of this list). Possible pre-specified themes are \code{vintage} (default), \code{colored}, \code{delftware}, \code{lichtenstein}, \code{original}, \code{rouge} and \code{modern}.}

\item{border}{a character specifying the type of border to use. Possible options are \code{none} (default), \code{circle}, \code{rhombus}, \code{square}, \code{hexagon} (6 vertices), \code{octagon} (8 vertices) and \code{decagon} (10 vertices).}

\item{halftone}{a character specifying the type of halftone to use. Possible options are \code{none}, \code{light} (white dither) and \code{dark} (black dither).}

\item{places}{an integer specifying how many suburb, quarter and neighbourhood names to add to the image.}

\item{legend}{logical. Whether to add a distance measurer and a compass in the bottom left corner of the image.}

\item{filename}{character. If specified, the function exports the plot at an appropriate size and does not return a \code{ggplot2} object.}

\item{verbose}{logical. Whether to show a progress bar during execution.}

\item{license}{logical. Whether to add the OpenStreetMap licence to the plot.}

\item{bot}{logical. Enable functionality used by the Twitter bot.}
}
\description{
Create an aerial city view.
}
\details{
The \code{theme} argument can take a custom list as input. This list must contain the following elements:

\code{colors}
\itemize{
 \item{\code{background}:  A color for the background.}
 \item{\code{water}:       A color for the water.}
 \item{\code{landuse}:     A color or vector of colors for the landuse.}
 \item{\code{contours}:    A color for the contours of landuse and buildings.}
 \item{\code{streets}:     A color for the streets.}
 \item{\code{buildings}:   A color or vector of colors for the buildings.}
 \item{\code{text}:        A color for the text.}
}
\code{font}
\itemize{
 \item{\code{family}:  the family of the font}
 \item{\code{face}:    the face of the font.}
 \item{\code{append}:      Optional. A string to append the city name at both sides.}
}
\code{size}
\itemize{
 \item{\code{borders}:  the family of the font}
 \item{\code{streets}:    the face of the font.}
}
}
\examples{
\dontrun{
# Create a vintage city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "vintage", border = "circle")

# Use a custom theme
theme <- list(
  colors = list(
    "background" = "#a25543",
    "water" = "#f2deb8",
    "landuse" = "#a25543",
    "contours" = "#f2deb8",
    "streets" = "#f2deb8",
    "buildings" = "#a25543",
    "text" = "#f2deb8",
    "waterlines" = "#f2deb8"
  ),
  font = list(
    "family" = "Oswald",
    "face" = "bold"
  ),
  size = list(
    "borders" = list(
      "contours" = 0.3,
      "water" = 0.4,
      "canal" = 0.5,
      "river" = 0.6
    ),
    "streets" = list(
      "path" = 0.1,
      "residential" = 0.4,
      "structure" = 0.5,
      "tertiary" = 0.75,
      "secondary" = 0.8,
      "highway" = 0.9,
      "motorway" = 1,
      "rails" = 0.75,
      "runway" = 3
    )
  )
)
cityview(name = "Amsterdam", theme = theme)
}
}
\seealso{
\code{\link{list_cities}} \code{\link{cityview_shiny}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
