% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview.R
\name{cityview}
\alias{cityview}
\title{Create an Aerial City View}
\usage{
cityview(name = NULL,
         zoom = 1,
         theme = c(
           "original", "light", "dark", "colored", "rouge",
           "verde", "neon", "delftware", "vintage", "lichtenstein"
         ),
         border = c(
           "none", "circle", "rhombus", "square",
           "hexagon", "octagon", "decagon"
         ),
         halftone = c("none", "light", "dark"),
         places = 0,
         legend = FALSE,
         filename = NULL,
         verbose = TRUE,
         license = TRUE,
         bot = FALSE)
}
\arguments{
\item{name}{a character specifying the name of the city as provided by \code{list_cities()}. If \code{NULL} (default), chooses a random city.}

\item{zoom}{a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.}

\item{theme}{a character specifying the theme of the plot, or a list of colors. See the details section for more information. Possible options are \code{original}, \code{light}, \code{dark}, \code{colored}, \code{rouge}, \code{verde}, \code{neon}, \code{delftware}, \code{vintage} and \code{lichtenstein}.}

\item{border}{a character specifying the type of border to use. Possible options are \code{none} (default), \code{circle}, \code{rhombus}, \code{square}, \code{hexagon} (6 vertices), \code{octagon} (8 vertices) and \code{decagon} (10 vertices).}

\item{halftone}{a character specifying the type of halftone to use. Possible options are \code{none}, \code{light} (white dither) and \code{dark} (black dither).}

\item{places}{an integer specifying how many suburb, quarter and neighbourhood names to add to the image.}

\item{legend}{logical. Whether to add a distance measurer and a compass in the bottom left corner of the image.}

\item{filename}{character. If specified, the function exports the plot at an appropriate size and does not return a \code{ggplot2} object.}

\item{verbose}{logical. Whether to show a progress bar during execution.}

\item{license}{logical. Whether to add the OpenStreetMap licence to the plot.}

\item{bot}{logical. Enable functionality used by the Twitter bot.}
}
\description{
Create an aerial city view.
}
\details{
The \code{theme} argument can take a custom list as input. This list must contain the following elements:
}
\examples{
\dontrun{
# Create a city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "original", border = "circle")
}
}
\seealso{
\code{\link{list_cities}} \code{\link{cityview_shiny}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
