% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview.R
\name{cityview}
\alias{cityview}
\title{Create an Aerial City View}
\usage{
cityview(name = NULL,
         zoom = 1,
         theme = c(
           "vintage", "modern", "bright", "delftware",
           "comic", "rouge", "original"
         ),
         border = c(
           "none", "circle", "rhombus", "square",
           "hexagon", "octagon", "decagon"
         ),
         halftone = NULL,
         legend = FALSE,
         places = 0,
         license = TRUE,
         timeout = 25,
         filename = NULL,
         verbose = TRUE,
         bot = FALSE)
}
\arguments{
\item{name}{a character specifying the name of the city as provided by \code{list_cities()}. If \code{NULL} (default), chooses a random city.}

\item{zoom}{a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.}

\item{theme}{a character specifying the theme of the plot, or a named list specifying a custom theme (see the details section for more information about the composition of this list). Possible pre-specified themes are \code{vintage} (default), \code{modern}, \code{bright}, \code{delftware}, \code{comic}, \code{rouge} and \code{original}.}

\item{border}{a character specifying the type of border to use. Possible options are \code{none} (default), \code{circle}, \code{rhombus}, \code{square}, \code{hexagon} (6 vertices), \code{octagon} (8 vertices) and \code{decagon} (10 vertices).}

\item{halftone}{a character specifying the color of halftone to use. Possible options are \code{none}, \code{light} (white dither) and \code{dark} (black dither).}

\item{legend}{logical. Whether to add a distance measurer and a compass in the bottom left corner of the image.}

\item{places}{an integer specifying how many suburb, quarter and neighbourhood names to add to the image.}

\item{license}{logical. Whether to add the OpenStreetMap licence to the plot.}

\item{timeout}{a value specifying the timeout (seconds) for the overpass server. It may be necessary to increase this value for large or populated areas because the server may time out before all data are delivered (If this happends you will see a message: \code{runtime error: Query timed out in "recurse" at line ... after ... seconds}.}

\item{filename}{character. If specified, the function exports the plot at an appropriate size and does not return a \code{ggplot2} object.}

\item{verbose}{logical. Whether to show a progress bar during execution.}

\item{bot}{logical. Enable functionality used by the Twitter bot.}
}
\description{
Create an aerial city view.
}
\details{
The \code{theme} argument can take a custom list as input (see the example). This list must contain the following elements:

\code{colors}
\itemize{
 \item{\code{background}:  One color for the background.}
 \item{\code{water}:       One color for the water.}
 \item{\code{landuse}:     One color or a vector of multiple colors for the landuse.}
 \item{\code{contours}:    One color for the contours of landuse and buildings.}
 \item{\code{streets}:     One color for the streets.}
 \item{\code{rails}:       One color or a vector of two colors for the rails.}
 \item{\code{buildings}:   One color or a vector of multiple colors for the buildings.}
 \item{\code{text}:        One color for the text.}
}
\code{font}
\itemize{
 \item{\code{family}:      The family of the font.}
 \item{\code{face}:        The face of the font.}
 \item{\code{append}:      Optional. A string to append the city name with at both sides.}
}
\code{size}
\itemize{
 \item{\code{borders}:    A named list containing sizes for the borders \code{contours}, \code{water}, \code{canal} and \code{river}.}
 \item{\code{streets}:    A named list contianing sizes for the streets \code{path}, \code{residential}, \code{structure}, \code{tertiary}, \code{secondary}, \code{primary}, \code{motorway}, \code{rails} and \code{runway}.}
}
}
\examples{
\dontrun{
# Create a city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "vintage",
         border = "circle", filename = "Amsterdam.png")

# Use a custom theme, for example green and white, streets only
myTheme <- list(
  colors = list(
    background = "forestgreen",
    water = NA,
    landuse = NA,
    contours = NA,
    streets = "white",
    rails = c("white", "forestgreen"),
    buildings = NA,
    text = "white",
    waterlines = NA
  ),
  font = list(
    family = "serif",
    face = "bold",
   append = "\u2014"
  ),
  size = list(
    borders = list(
      contours = 0.3,
      water = 0.4,
      canal = 0.5,
      river = 0.6
    ),
    streets = list(
      path = 0.2,
      residential = 0.4,
      structure = 0.5,
      tertiary = 0.7,
      secondary = 0.8,
      primary = 0.9,
      motorway = 1,
      rails = 0.75,
      runway = 3
    )
  )
)
cityview(name = "Amsterdam", theme = myTheme,
         border = "square", filename = "Amsterdam.png")
}
}
\seealso{
\code{\link{list_cities}} \code{\link{cityview_shiny}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
