% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview.R
\name{cityview}
\alias{cityview}
\title{Create an Aerial City View}
\usage{
cityview(name, 
         zoom = 1,
         theme = c("original", "light", "dark", "colored", "rouge",
                   "verde", "neon", "atlantis", "vintage", "lichtenstein"),
         border = c("none", "circle", "rhombus", "square",
                    "hexagon", "octagon", "decagon"),
         halftone = c("none", "light", "dark", "auto"),
         places = FALSE, 
         filename = NULL, 
         verbose = TRUE,
         license = TRUE, 
         bot = FALSE)
}
\arguments{
\item{name}{a character specifying the name of the city as provided by \code{list_cities()}.}

\item{zoom}{a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.}

\item{theme}{a character specifying the theme of the plot. Possible options are \code{original}, \code{light}, \code{dark}, \code{colored}, \code{rouge}, \code{verde}, \code{neon}, \code{atlantis}, \code{vintage} and \code{lichtenstein}.}

\item{border}{a character specifying the type of border to use. Possible options are \code{none}, \code{circle}, \code{rhombus}, \code{square}, \code{hexagon} (6 vertices), \code{octagon} (8 vertices), and \code{decagon} (10 vertices).}

\item{halftone}{a character specifying the type of halftone to use. Possible options are \code{none}, \code{light} (white dither), \code{dark} (black dither) and \code{auto} (color according to theme).}

\item{places}{logical. Whether to add neighborhood names to the plot.}

\item{filename}{character. If specified, the function exports the plot at an appropriate size and does NOT return a \code{ggplot2} object.}

\item{verbose}{logical. Whether to show a progress bar during execution.}

\item{license}{logical. Whether to add the OpenStreetMap licence to the plot.}

\item{bot}{logical. Choose automatically between cities with the same name and add a copyright licence to the image. Primarily used by the twitter bot.}
}
\description{
Create an aerial city view.
}
\examples{
\dontrun{
# Create a city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "original", border = "circle")
}
}
\seealso{
\code{\link{list_cities}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
