% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityview_shiny.R
\name{cityview_shiny}
\alias{cityview_shiny}
\title{[Shiny] Create a City View}
\usage{
cityview_shiny(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{logical. If \code{FALSE} (the default), the app will
open in the viewer panel. If \code{TRUE}, the app will
open in the system's default web browser.}
}
\description{
Launch a Shiny app in which you can create a city view
  showcasing a particular city or region using OpenStreetMap (OSM) data
  retreived trough the Overpass API. Please note that OpenStreetMap is open
  data and can be freely utilized for any purpose, as long as proper credit
  is given to OpenStreetMap and its contributors.
}
\examples{
\dontrun{
# Start the shiny app
cityview_shiny()
}
}
\references{
\url{https://www.openstreetmap.org}
}
\seealso{
\code{\link{list_cities}}
         \code{\link{cityview}}
         \code{\link{new_city}}
         \code{\link{city_themes}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{create}
\keyword{shiny}
