% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cities.R
\name{list_cities}
\alias{list_cities}
\title{List or Search City Names}
\usage{
list_cities(match = NULL)
}
\arguments{
\item{match}{string to match.}
}
\value{
A data frame containing all matched cities alongside their respective
  country and coordinates. A row of the output can be used as input for the
  \code{name} argument in the \code{cityview()} function.
}
\description{
List all city names containing a specific string.
}
\examples{
# List all cities containing "Ams"
list_cities(match = "Ams")
}
\seealso{
\code{\link{cityview}}
         \code{\link{cityview_shiny}}
         \code{\link{new_city}}
         \code{\link{city_themes}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{search}
