% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_city.R
\name{new_city}
\alias{new_city}
\title{Specify a New City}
\usage{
new_city(name = NULL, country = NULL, lat = NULL, long = NULL)
}
\arguments{
\item{name}{A single string to be used as the city name.}

\item{country}{A single string to be used as the country.}

\item{lat}{A single numeric value to be used as the latitude.}

\item{long}{A single numeric value to be used as the longitude.}

\item{method}{A character string specifying the geocoding method to use when the user does not specify \code{lat} or \code{long}. Defaults to "osm" (OpenStreetMap). Supported methods include:
\code{"osm"}, \code{"census"}, \code{"arcgis"}, \code{"census_simple"}, \code{"geocodio"}, \code{"mapbox"},
\code{"google"}, \code{"bing"}, \code{"here"}, \code{"tomtom"}, \code{"nominatim"}, and \code{"tiger"}. Visit \url{https://jessecambon.github.io/tidygeocoder/reference/geo.html} for more details}
}
\value{
a data frame containing the new city alongside its respective
  country and coordinates.
}
\description{
Define a city using country, latitude and longitude coordinates.
  The returned object can be used a input for \code{name} argument the
  \code{cityview()} function.
}
\examples{
city <- new_city(
  name = "Lagos", country = "Portugal",
  lat = 37.10, long = -8.68
)
\dontrun{
cityview(name = city)
}
}
\seealso{
\code{\link{cityview}}
         \code{\link{cityview_shiny}}
         \code{\link{list_cities}}
         \code{\link{city_themes}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{cities}
\keyword{new}
