% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcityviews-methods.R
\name{rcityviews-methods}
\alias{rcityviews-methods}
\alias{print.rcityviewsCity}
\alias{print.rcityviewsTheme}
\title{Methods for rcityviews Objects}
\usage{
\method{print}{rcityviewsCity}(x, ...)

\method{print}{rcityviewsTheme}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rcityviewsCity} or
\code{rcityviewsTheme}.}

\item{...}{further arguments, currently ignored.}
}
\value{
The \code{print} methods simply print and return nothing.
}
\description{
Methods defined for objects returned from the \code{\link{new_city}} and
  \code{\link{city_themes}} functions.
}
